Replaced initialization with TUI object
