+++ Makefile
@@ -1,18 +1,21 @@
 CC=g++
-CFLAGS = -g -std=c++11 -Wall
+CFLAGS = -g -std=c++17 -Wall
 
 LIBS = -lncurses
 
-clock: main.o clock.o clockMenu.o frame.o digit.o
-	$(CC) -o clock main.o clock.o clockMenu.o frame.o digit.o $(CFLAGS) $(LIBS)
+clock: main.o TUI.o clock.o clockMenu.o frame.o digit.o
+	$(CC) -o clock main.o TUI.o clock.o clockMenu.o frame.o digit.o $(CFLAGS) $(LIBS)
 	rm *.o
 
-main.o: src/main.cpp src/clock.hpp src/clockMenu.hpp
+main.o: src/main.cpp src/TUI.hpp src/clock.hpp src/clockMenu.hpp
 	$(CC) -c src/main.cpp $(CFLAGS)
+
+TUI.o: src/TUI.cpp src/TUI.hpp
+	$(CC) -c src/TUI.cpp $(CFLAGS)
 
 clock.o: src/clock.cpp src/clock.hpp
 	$(CC) -c src/clock.cpp $(CFLAGS)
- 
+
 clockMenu.o: src/clockMenu.cpp src/clockMenu.hpp
 	$(CC) -c src/clockMenu.cpp $(CFLAGS)
 
@@ -24,4 +27,3 @@ digit.o: src/digit.cpp src/digit.hpp
 
 clean:
 	rm clock
-
