+++ src/TUI.hpp
@@ -7,6 +7,9 @@ class TUI
   private:
 
   public:
+    TUI();
+    ~TUI();
+
     void DisplayMessageAt(const std::string& message, int x, int y);
 };
