+++ src/main.cpp
@@ -1,3 +1,4 @@
+#include "TUI.hpp"
 #include "clock.hpp"
 #include "clockMenu.hpp"
 #include <chrono>
@@ -12,18 +13,6 @@
  */
 int main()
 {
-    enum ColorPairs
-    {
-        PAIR_BLACK = 1,
-        PAIR_RED = 2,
-        PAIR_GREEN = 3,
-        PAIR_YELLOW = 4,
-        PAIR_BLUE = 5,
-        PAIR_MAGENTA = 6,
-        PAIR_CYAN = 7,
-        PAIR_WHITE = 8
-    };
-
     const int MENU_COLOR_PAIR = 9;
 
     bool readClock = false, readDigit = false;
@@ -33,7 +22,7 @@ int main()
     {
 
         std::string line;
-        
+
         while(std::getline(inFile, line))
         {
             std::stringstream lineStream(line);
@@ -60,34 +49,7 @@ int main()
         }
     }
 
-    initscr();
-    keypad(stdscr, TRUE);
-    // Don't pause for user input.
-    nodelay(stdscr, TRUE); 
-    // Don't echo usr input.
-    noecho();
-    // Turn off cursor.
-    curs_set(0);
-
-    if(has_colors() == false)
-    {
-        endwin();
-        printf("Color is not supported by terminal.");
-        exit(1);
-    }
-
-    // Set up colors.
-    start_color();
-    use_default_colors(); 
-    init_pair(PAIR_BLACK, COLOR_BLACK, -1);
-    init_pair(PAIR_RED, COLOR_RED, -1);
-    init_pair(PAIR_GREEN, COLOR_GREEN, -1);
-    init_pair(PAIR_YELLOW, COLOR_YELLOW, -1);
-    init_pair(PAIR_BLUE, COLOR_BLUE, -1);
-    init_pair(PAIR_MAGENTA, COLOR_MAGENTA, -1);
-    init_pair(PAIR_CYAN, COLOR_CYAN, -1);
-    init_pair(PAIR_WHITE, COLOR_WHITE, -1);
-    init_pair(MENU_COLOR_PAIR, COLOR_WHITE, COLOR_BLUE);
+    TUI tui;
 
     // Tracks time passed for updates.
     std::chrono::milliseconds now = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch());
@@ -114,7 +76,7 @@ int main()
         using namespace std::chrono;
         now = duration_cast<milliseconds>(system_clock::now().time_since_epoch());
         milliseconds elapsed = now - prev;
-        
+
         if(elapsed > milliseconds(50))
         {
             clear();
@@ -123,7 +85,7 @@ int main()
             {
                 menu.displayMenu();
             }
-            
+
             refresh();
             prev = now;
         }
@@ -188,6 +150,5 @@ int main()
         }
     }
 
-    endwin();
     return 0;
 }
